+++ app/components/omnibox/directives/search-directive.js
@@ -96,6 +96,11 @@ angular.module('omnibox')
       UtilService.announceMovedTimeline(State);
     };
 
+
+    var prevKey; // stores the previously pressed key.
+    var prevKeyTimeout; // resets the key after TIMEOUT.
+    var TIMEOUT = 300; // 300 ms
+
     /**
      * @description event handler for key presses.
      * checks if enter is pressed, does search.
@@ -103,7 +108,9 @@ angular.module('omnibox')
      * 13 refers to the RETURN key.
      */
     scope.searchKeyPress = function ($event) {
+      clearTimeout(prevKeyTimeout);
       var KEYPRESS = {
+        BACKSPACE: 8,
         ENTER: 13,
         SPACE: 32,
         ESC: 27
@@ -112,7 +119,16 @@ angular.module('omnibox')
       if ($event.target.id === "searchboxinput") {
         // Intercept keyPresses *within* searchbox,do xor prevent animation
         // from happening when typing.
-        if ($event.which === KEYPRESS.ENTER) {
+        if ($event.which === KEYPRESS.ESC) {
+          scope.cleanInput();
+        } else if ($event.which === KEYPRESS.BACKSPACE && prevKey === KEYPRESS.BACKSPACE) {
+          scope.omnibox.searchResults = {}; // only delete search results
+        } else if ($event.which === KEYPRESS.BACKSPACE && scope.query === "") {
+          scope.omnibox.searchResults = {}; // only delete search results           
+        } else if ($event.which === KEYPRESS.SPACE) {
+          // prevent anim. start/stop
+          $event.originalEvent.stopPropagation();
+        } else if ($event.which === KEYPRESS.ENTER) {
           var loc = scope.omnibox.searchResults;
           if (loc && loc.temporal) {
             scope.zoomToTemporalResult(
@@ -127,13 +143,14 @@ angular.module('omnibox')
           else {
             scope.search();
           }
-        } else if ($event.which === KEYPRESS.SPACE) {
-          // prevent anim. start/stop
-          $event.originalEvent.stopPropagation();
-        } else if ($event.which === KEYPRESS.ESC) { //esc
-          scope.cleanInput();
+        } else {
+          scope.search();
         }
       }
+      prevKey = $event.which;
+      prevKeyTimeout = setTimeout(function () {
+        prevKey = null;
+      }, TIMEOUT);
     };
 
     /**